src/main/java/com/chifungcheung/leetcode/solutions/s84/Solution.java
package com.chifungcheung.leetcode.solutions.s84;

import java.util.Arrays;
import java.util.Stack;

/*
 * @lc app=leetcode id=84 lang=java
 *
 * [84] Largest Rectangle in Histogram
 */

// @lc code=start
class Solution {
    public int largestRectangleArea(int[] heights) {
        Stack<Integer> stack = new Stack<>();
        int maxArea = 0;
        heights = Arrays.copyOf(heights, heights.length + 1);
        heights[heights.length - 1] = 0;
        for (int i = 0; i < heights.length; i++) {
            while (!stack.empty() && heights[stack.peek()] >= heights[i]) {
                int c = stack.pop();
                int h = heights[c];
                int idx = stack.empty() ? -1 : stack.peek();
                int w = i - idx - 1;
                maxArea = Math.max(maxArea, h * w);
            }
            stack.push(i);
        }
        return maxArea;
    }
}
// @lc code=end
